Separate a common part of MongoDBStandaloneLauncher.
Some part of MongoDBStandaloneLauncher can be used for other MongoDB
instances such as shards or config servers. This patch separates the
part into MongoDBLauncher.
